requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
python-dotenv==1.0.1
firebase-admin==6.5.0
pydantic==2.7.1
